Added some more splits
Fire Arrows, Spooky Mask, Sell Bunny Hood, Mask of Truth, 2nd Stick
Upgrade, Deku Nuts Upgrade, Adult 2, Scrubs and RichardZ.

+++ BombBag.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace LiveSplit.OcarinaOfTime
 {
-    public enum BombBag : ushort
+    public enum BombBag : uint
     {
         None = 0x0,
         BombBag = 0x1,
